New project to push to GitHub

+++ app/Http/Controllers/PostController.php
@@ -1,10 +1,21 @@
 <?php
 
 namespace App\Http\Controllers;
+
+use App\Models\Category;
+use App\Models\Comment;
 use DB;
 use App\Models\Post;
+use App\Models\Tag;
+use App\Models\User;
+use Illuminate\Auth\Middleware\Authorize;
 use Illuminate\Http\Request;
+use Illuminate\Contracts\Auth\Authenticatable;
+use Illuminate\Foundation\Auth\User as AuthUser;
+use Illuminate\Support\Facades\Auth;
 use PhpParser\Model\Stmt\Return_;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Support\Facades\Validator as FacadesValidator;
 
 class PostController extends Controller
 {
@@ -13,8 +24,9 @@ class PostController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAny',Post::class);
         $posts= Post::all();
-        return view("posts.index")->with("posts",$posts);
+        return view("posts.index",compact('posts'));
     }
 
     /**
@@ -22,29 +34,45 @@ class PostController extends Controller
      */
     public function create()
     {
-        return view("posts.create");
+        $this->authorize('create',Post::class);
+        $tag= Tag::all();
+        $category= Category::all();
+        return view("posts.create",['tag' => $tag,'category'=>$category]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
-    {   $tags=implode(",",$request->tags);
-        $post=new Post;
-        if($request->hasfile('image')){
-                $img=$request['image'];
-                $imgname=time().".".$img->getClientOriginalExtension();
-                $img->move('./assets/imgs',$imgname);
+    {
+        $validator = FacadesValidator::make($request->all(),[
+        'title'=>'required|max:50',
+        'description'=>'required|string',
+        'image'=>'image|mimes:jpeg,jpg,png,gif|max:2048|required',
+        'category'=>'required',
+        'tag'=>'required|array',
+        'user_id'=>'require',
+        ]);
+        if($validator->fails()){
+            return response($validator->messages(), 200);
         }
-        $post->create(
+        $post=new Post;
+        if($request->hasFile('image')){
+                $img=$request['image'];
+                $imgName=time().".".$img->getClientOriginalExtension();
+                $img->move('./assets/imgs',$imgName);
+                $data['image']=$imgName;
+        }
+        $post1=$post->create(
         [  'title'=>$request['title'],
             'description'=>$request['description'],
-            'tags'=>$tags,
-            'show'=>$request->show,
-            'categories'=>$request->category,
-            'image'=>$imgname
+            'image'=>$imgName,
+            'category_id'=>$request->category,
+            'user_id' => Auth::id(),
         ]);
-        return redirect()->route('post.index')->with("success","post add successfuly");
+        /* $posts= Post::where(['title' => $request->title, 'description' => $request->description])->first()->id; */
+        $post1->tags()->attach($request->tag);
+        return redirect()->route('post.index')->with("success","post add successfully");
     }
 
     /**
@@ -52,16 +80,24 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
-        /* $post= DB::table("posts")->where("id",$id)->first(); */
-        return view("posts.show")->with("post",$post);
+        $this->authorize('view',$post);
+        $user=Auth::user();
+        $tags=$post->tags()->get();
+        $comment = Comment::where('post_id', $post->id)->get();
+        return view("posts.show",compact('post','user' , 'tags','comment'));
     }
 
     /**
      * Show the form for editing the specified resource.
      */
     public function edit(Post $post)
-    {
-        return view("posts.edit")->with("post",$post);
+
+    {   $this->authorize('update',$post);
+        $tag= Tag::all();
+        $tags=$post->tags()->get();
+        $category= Category::all();
+        $user=Auth::user();
+        return view("posts.edit",['tag'=>$tag,'category'=>$category,'user'=>$user,'post'=>$post,'tags'=>$tags]);
     }
 
     /**
@@ -69,22 +105,33 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
-        $imgname=$post->image;
-        if($request->hasfile('image')){
-        $img=$request['image'];
-        $imgname=time().".".$img->getClientOriginalExtension();
-        $img->move('./assets/imgs',$imgname);
-        }
-        $tags=implode(",",$request->tags);
-        $post->update([
-                    'title'=>$request['title'],
+            $validator = FacadesValidator::make($request->all(),[
+            'title'=>'required|max:50',
+            'description'=>'required|string',
+            'image'=>'image|mimes:jpeg,jpg,png,gif|max:2048|nullable',
+            'category'=>'required',
+            'tag'=>'required|array',
+            ]);
+            if($validator->fails()){
+                return response($validator->messages(), 200);
+            }
+            $imgName=$post->image;
+            if($request->hasFile('image')){
+                $img=$request['image'];
+                $imgName=time().".".$img->getClientOriginalExtension();
+                $img->move('./assets/imgs',$imgName);
+                $data['image']=$imgName;
+                }
+                $post->update(
+                [  'title'=>$request['title'],
                     'description'=>$request['description'],
-                    'tags'=>$tags,
-                    'show'=>$request->show,
-                    'categories'=>$request->category,
-                    'image'=>$imgname
-        ]);
-        return redirect()->route('post.index')->with("success","post add successfuly");
+                    'image'=>$imgName,
+                    'user_id' => Auth::id(),
+                    'category_id'=>$request->category,
+                ]);
+                $post->tags()->detach($post->tags()->get());
+                $post->tags()->attach($request->tag) ;
+                return redirect()->route('post.index')->with("success","post updated successfully");
 
     }
 
@@ -93,6 +140,7 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
 {
+    $this->authorize('delete',$post);
     $post->delete();
-return redirect()->route('post.index')->with("success","post add successfuly");}
+    return redirect()->route('post.index')->with("success","post deleted successfully");}
 }
